+++ packages/api/lib/granules.js
@@ -209,16 +209,35 @@ async function moveGranule(apiGranule, destinations, distEndpoint, granulesModel
 const SCROLL_SIZE = 500; // default size in Kibana
 const TIMEOUT = '60s';
 
+function getTotalHits(bodyHits) {
+  if (bodyHits.total.value === 0) {
+    return 0;
+  }
+  return bodyHits.total.value || bodyHits.total;
+}
+
+/**
+ * Returns an array of granules from ElasticSearch query
+ *
+ * @param {Object} payload
+ * @param {string} [payload.index] - ES index to query
+ * @param {string} [payload.query] - ES query
+ * @param {Object} [payload.source] - List of IDs to operate on
+ * @param {Object} [payload.timeout] - Optional timeout variable for ES search operation.
+ *   Defaults to 60 seconds.
+ * @param {Object} [payload.testBodyHits] - Optional body.hits for testing
+ * @returns {Promise<Array<Object>>}
+ */
 async function granuleEsQuery({
   index,
   query,
   source,
   timeout,
+  testBodyHits,
 }) {
   const granules = [];
   const responseQueue = [];
   const searchTimeout = timeout || TIMEOUT;
-  log.info(`Timeout for search operation is ${searchTimeout}`);
 
   const client = await Search.es(undefined, true);
   const searchResponse = await client.search({
@@ -232,14 +251,16 @@ async function granuleEsQuery({
 
   responseQueue.push(searchResponse);
 
-  for (let i = 0; i < responseQueue.length; i++) {
+  while (responseQueue.length) {
     const { body } = responseQueue.shift();
+    const bodyHits = testBodyHits || body.hits;
 
-    body.hits.hits.forEach((hit) => {
+    bodyHits.hits.forEach((hit) => {
       granules.push(hit._source);
     });
 
-    const totalHits = body.hits.total.value || body.hits.total;
+    const totalHits = getTotalHits(bodyHits);
+    console.log(totalHits);
 
     if (totalHits !== granules.length) {
       responseQueue.push(